+++ HappyTravel.Edo.Common/HappyTravel.Edo.Common.csproj
@@ -9,8 +9,4 @@
         <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
     </ItemGroup>
 
-    <ItemGroup>
-      <Folder Include="Enums\Enums" />
-    </ItemGroup>
-
 </Project>
