+++ HappyTravel.Edo.NotificationCenter/Services/Notification/INotificationService.cs
@@ -1,12 +1,13 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using HappyTravel.Edo.Common.Enums.Notifications;
 using HappyTravel.Edo.NotificationCenter.Models;
 
-namespace HappyTravel.Edo.NotificationCenter.Services.Message
+namespace HappyTravel.Edo.NotificationCenter.Services.Notification
 {
     public interface INotificationService
     {
-        Task Add(Notification notification);
+        Task Add(Models.Notification notification);
         Task MarkAsRead(int notificationId);
         Task<List<NotificationSlim>> GetMessages(int userId, int top, int skip);
     }
